Consolidate sceneFilterPane component
