--- app-frontend/src/app/components/scenes/sceneFilterPane/sceneFilterPane.component.js
@@ -1,15 +1,0 @@
-import sceneFilterTpl from './sceneFilterPane.html';
-
-const rfSceneFilterPane = {
-    templateUrl: sceneFilterTpl,
-    controller: 'SceneFilterPaneController',
-    bindings: {
-        filters: '<',
-        opened: '<',
-        onFilterChange: '&',
-        onCloseFilterPane: '&',
-        onPassPlanetToken: '&'
-    }
-};
-
-export default rfSceneFilterPane;
